Added method to calculate rapid attacks.
Also fixed bug in method for simple attacks, where the attack dice
instead of the damage dice were being updated.

Additionally, introduced some branches at the beginning to decide
which works to run. But I am not happy with so many branches (and more
to come, when we add rapid attacks and maybe other options); there
must be a more elegant way.