I have abandoned the idea of simulation results that include the simulation of different options (wild, non-wild, head attack, normal attack, etc) because it makes the code too complex for very little gain. I have returned to the idea of simulating just one case, and counting the wounds for only that case (i.e. using DamageCounters instead of SimulationResults).
The whole idea of SimulationResults was thought to facilitate a
recommender system of tactical options for SavageWorlds, but I think
this can be more easily implemented in a different way, i.e. a
different frame (not the simulator frame) that launches several
simulation jobs (each of them returning a DamageCounter), collects the
results, and then shows the recommendation. There is no need for this
mess of simulation results, nested ifs in the simulator, etc.